feat: move remove_user_from_affiliated_communities_and_institutions function to accounts/utils.py
